backend/api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
asyncpg==0.29.0
pgvector==0.2.4
litellm==1.18.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
# For SSE
sse-starlette==2.0.0
# For Realtime Feed
aiokafka==0.10.0
protobuf>=5.29.0
grpcio-tools>=1.60.0
websockets>=12.0
# For Mission Area Config
redis>=5.0.0
sgp4>=2.22
numpy>=1.26
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
